Correctly take requested reviews into account
After a bunch of testing I unleashed Octobors on one of our repos today,
and oine of the first things it did was to merge a PR that was waiting
for requested reviewers. Oh dear...

It turns out I made a mistake in detection of these requests. Rather
than being in the reviews field with a `Pending` state there is a
requested reviews field that I had missed! This PR takes this field into
account, resolving the problem.