Don't count comments post-approval as requests for changes (#29)
This has bitten us a few times in the past that a comment posted to a
thread was counted as a request for change, with the new setting. This
tweaks the behavior so it's not the case anymore: after a reviewer has
approved a PR, they can comment freely, but if they want to ask for
changes, they should be explicit about it and do a "Request changes"
review.