Add a trivial review label (#35)
* Add a trivial review label (fixes #34)

This adds a new trivial review label to the configuration, which if it
is present, makes it so that the PR doesn't block on reviews. Other
preconditions may still prevail (e.g. CI must pass, someone requested changes in the meanwhile), but the requirement
on reviewers in particular is relaxed.

Fixes #34.