Maintain most spacing and paragraph structure in commit messages (#41)
Does what it says on the tin. I've added tests that didn't pass before this change, and had to tweak a few tests, so one can easily look at the changes. This shouldn't badly interact with Markdown in general, unless we have very whitespace-specific content in code blocks that seems fine to have.

Fixes #33.